new: [shlib] support of autoloading.
This will allow other library to implement some code that will
be loaded in all shell-script. If this can be badly used, it can
also permit to define new syntax for kal-shlib scripts.

For instance, decorators can be implemented and off-load all the
'include' and 'depends' checks on first usage of concerned functions.